+++ ghost/core/core/server/lib/image/blog-icon.js
