+++ ghost/core/core/server/lib/image/cached-image-size-from-url.js
