+++ ghost/core/core/server/lib/image/gravatar.js
