+++ ghost/core/core/server/lib/image/image-size.js
