+++ ghost/core/core/server/services/custom-theme-settings.js
@@ -1,4 +1,4 @@
-const CustomThemeSettingsService = require('../../shared/custom-theme-settings-cache/CustomThemeSettingsService');
+const CustomThemeSettingsService = require('../../shared/custom-theme-settings-cache/custom-theme-settings-service');
 const customThemeSettingsCache = require('../../shared/custom-theme-settings-cache');
 const models = require('../models');
