+++ ghost/core/core/shared/custom-theme-settings-cache/index.js
@@ -1,3 +1,3 @@
-const CustomThemeSettingsCache = require('./CustomThemeSettingsCache');
+const CustomThemeSettingsCache = require('./custom-theme-settings-cache');
 
 module.exports = new CustomThemeSettingsCache();
