+++ ghost/core/core/shared/events-ts/post-deleted-event.ts
