+++ ghost/core/core/shared/events-ts/posts-bulk-add-tags-event.ts
