+++ ghost/core/core/shared/events-ts/posts-bulk-destroyed-event.ts
