+++ ghost/core/core/shared/events-ts/posts-bulk-featured-event.ts
