+++ ghost/core/core/shared/events-ts/posts-bulk-unpublished-event.ts
