+++ ghost/core/core/shared/events/index.js
@@ -1,17 +1,17 @@
 module.exports = {
-    MemberCommentEvent: require('./MemberCommentEvent'),
-    MemberCreatedEvent: require('./MemberCreatedEvent'),
-    MemberEntryViewEvent: require('./MemberEntryViewEvent'),
-    MemberLinkClickEvent: require('./MemberLinkClickEvent'),
-    MemberPageViewEvent: require('./MemberPageViewEvent'),
-    MemberPaidCancellationEvent: require('./MemberPaidCancellationEvent'),
-    MemberPaidConversionEvent: require('./MemberPaidConversionEvent'),
-    MemberSignupEvent: require('./MemberSignupEvent'),
-    MemberSubscribeEvent: require('./MemberSubscribeEvent'),
-    MemberUnsubscribeEvent: require('./MemberUnsubscribeEvent'),
-    OfferRedemptionEvent: require('./OfferRedemptionEvent'),
-    SubscriptionActivatedEvent: require('./SubscriptionActivatedEvent'),
-    SubscriptionCancelledEvent: require('./SubscriptionCancelledEvent'),
-    SubscriptionCreatedEvent: require('./SubscriptionCreatedEvent'),
-    URLResourceUpdatedEvent: require('./URLResourceUpdatedEvent')
+    MemberCommentEvent: require('./member-comment-event'),
+    MemberCreatedEvent: require('./member-created-event'),
+    MemberEntryViewEvent: require('./member-entry-view-event'),
+    MemberLinkClickEvent: require('./member-link-click-event'),
+    MemberPageViewEvent: require('./member-page-view-event'),
+    MemberPaidCancellationEvent: require('./member-paid-cancellation-event'),
+    MemberPaidConversionEvent: require('./member-paid-conversion-event'),
+    MemberSignupEvent: require('./member-signup-event'),
+    MemberSubscribeEvent: require('./member-subscribe-event'),
+    MemberUnsubscribeEvent: require('./member-unsubscribe-event'),
+    OfferRedemptionEvent: require('./offer-redemption-event'),
+    SubscriptionActivatedEvent: require('./subscription-activated-event'),
+    SubscriptionCancelledEvent: require('./subscription-cancelled-event'),
+    SubscriptionCreatedEvent: require('./subscription-created-event'),
+    URLResourceUpdatedEvent: require('./url-resource-updated-event')
 };
