+++ ghost/core/core/shared/events/member-created-event.js
