+++ ghost/core/core/shared/events/member-entry-view-event.js
