+++ ghost/core/core/shared/events/member-link-click-event.js
