+++ ghost/core/core/shared/events/member-page-view-event.js
