+++ ghost/core/core/shared/events/member-signup-event.js
