+++ ghost/core/core/shared/events/url-resource-updated-event.js
