+++ ghost/core/core/shared/settings-cache/cache-manager.js
