+++ ghost/core/core/shared/settings-cache/index.js
@@ -1,4 +1,4 @@
-const CacheManager = require('./CacheManager');
+const CacheManager = require('./cache-manager');
 const publicSettings = require('./public');
 
 const cacheManager = new CacheManager({publicSettings});
